refactor(financeiro): split FinanceiroCockpit (509→130) em hook + componentes
Extrai verbatim de src/pages/FinanceiroCockpit.tsx:
- useFinanceiroCockpit: state, loadAll (resumo/aging/DRE/inadimplentes/
  projeção 13s/confiabilidade) e todos os derivados consolidados
- CockpitHeader, CockpitCard, MiniCard, TransparencyBadge, ResultadoPorEmpresa,
  Projecao13Card, TopInadimplentes, DataBasisFooter, types.ts, format.ts
  (fmt/fmtCompact/IMPOSTO_LABEL)
- +15 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0 erros
(1 warning exhaustive-deps pré-existente, apenas relocado para o hook).
